[REFACTOR] #12 Travel 잔액 정산 타입 변경

+++ backend/src/main/java/com/ssafy/moyeobang/settle/application/domain/travel/Travel.java
@@ -4,7 +4,6 @@ import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
-import java.math.BigInteger;
 import java.util.HashMap;
 import java.util.Map;
 
@@ -16,9 +15,9 @@ public class Travel {
 
     private String title;
 
-    private Map<Long, BigInteger> memberBalanceMap;
+    private Map<Long, Integer> memberBalanceMap;
 
-    public static Travel of(Long id, String title, Map<Long, BigInteger> memberBalance) {
+    public static Travel of(Long id, String title, Map<Long, Integer> memberBalance) {
 
         return new Travel(id, title, memberBalance);
     }
